pyproject.toml
[project]
name = "paperops-template"
version = "0.1.0"
description = "PaperOps template project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  "psutil>=5.9.8",
  "snakemake>=8.0.0",
]

[project.scripts]
paperops = "paperops.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]

[tool.black]
line-length = 88

[tool.uv]
package = true
